+++ scout/adapter/mongo/hgnc.py
@@ -2,7 +2,6 @@ import logging
 from pprint import pprint as pp
 import intervaltree
 
-from scout.build.genes.exon import build_exon
 from pymongo.errors import (DuplicateKeyError, BulkWriteError)
 
 from scout.exceptions import IntegrityError
@@ -44,55 +43,6 @@ class GeneHandler(object):
 
         return result
 
-    def load_hgnc_transcript(self, transcript_obj):
-        """Add a transcript object to the database
-
-        Arguments:
-            transcript_obj(dict)
-
-        """
-        res = self.transcript_collection.insert_one(transcript_obj)
-        return res
-
-    def load_transcript_bulk(self, transcript_objs):
-        """Load a bulk of transcript objects to the database
-
-        Arguments:
-            transcript_objs(iterable(scout.models.hgnc_transcript))
-
-        """
-        LOG.info("Loading transcript bulk")
-        try:
-            result = self.transcript_collection.insert_many(transcript_objs)
-        except (DuplicateKeyError, BulkWriteError) as err:
-            raise IntegrityError(err)
-
-        return result
-
-    def load_exon(self, exon_obj):
-        """Add a exon object to the database
-
-        Arguments:
-            exon_obj(dict)
-
-        """
-        res = self.exon_collection.insert_one(exon_obj)
-        return res
-
-    def load_exon_bulk(self, exon_objs):
-        """Load a bulk of exon objects to the database
-
-        Arguments:
-            exon_objs(iterable(scout.models.hgnc_exon))
-
-        """
-        try:
-            result = self.exon_collection.insert_many(transcript_objs)
-        except (DuplicateKeyError, BulkWriteError) as err:
-            raise IntegrityError(err)
-
-        return result
-
     def hgnc_gene(self, hgnc_identifier, build='37'):
         """Fetch a hgnc gene
 
@@ -122,7 +72,8 @@ class GeneHandler(object):
         # Add the transcripts:
         transcripts = []
         tx_objs = self.transcripts(build=build, hgnc_id=gene_obj['hgnc_id'])
-        if tx_objs.count() > 0:
+        nr_tx = sum(1 for i in self.transcripts(build=build, hgnc_id=gene_obj['hgnc_id']))
+        if nr_tx > 0:
             for tx in tx_objs:
                 transcripts.append(tx)
         gene_obj['transcripts'] = transcripts
@@ -143,11 +94,10 @@ class GeneHandler(object):
         query = {'hgnc_symbol':hgnc_symbol, 'build':build}
         projection = {'hgnc_id':1, '_id':0}
         res = self.hgnc_collection.find(query, projection)
+        for gene in res:
+            return gene['hgnc_id']
 
-        if res.count() > 0:
-            return res[0]['hgnc_id']
-        else:
-            return None
+        return None
 
     def hgnc_genes(self, hgnc_symbol, build='37', search=False):
         """Fetch all hgnc genes that match a hgnc symbol
@@ -165,15 +115,16 @@ class GeneHandler(object):
         LOG.debug("Fetching genes with symbol %s" % hgnc_symbol)
         if search:
             # first search for a full match
-            full_query = self.hgnc_collection.find({
+            query = {
                 '$or': [
                     {'aliases': hgnc_symbol},
                     {'hgnc_id': int(hgnc_symbol) if hgnc_symbol.isdigit() else None},
                 ],
                 'build': build
-            })
-            if full_query.count() != 0:
-                return full_query
+            }
+            nr_genes = self.nr_genes(query=query)
+            if nr_genes != 0:
+                return self.hgnc_collection.find(query)
 
             return self.hgnc_collection.find({
                 'aliases': {'$regex': hgnc_symbol, '$options': 'i'},
@@ -182,32 +133,66 @@ class GeneHandler(object):
 
         return self.hgnc_collection.find({'build': build, 'aliases': hgnc_symbol})
 
-    def all_genes(self, build='37'):
+    def all_genes(self, build=None, add_transcripts=False, limit=100000):
         """Fetch all hgnc genes
-
+        
+        Args:
+            build(str)
+            add_transcripts(bool): If tx information should be added
+            limit(int): If only a part of the genes should be added
+            
+        
             Returns:
-                result()
+                genes(iterable):
+                limit(int): Maximum number of returned
         """
+        build = build or '37'
         if build == 'GRCh38':
             build = '38'
 
         LOG.info("Fetching all genes")
-        return self.hgnc_collection.find({'build': build}).sort('chromosome', 1)
+        
+        hgnc_tx = {}
+        if add_transcripts:
+            LOG.info("Adding transcripts")
+            for tx in self.transcripts(build=build):
+                hgnc_id = tx['hgnc_id']
+                if not hgnc_id in hgnc_tx:
+                    hgnc_tx[hgnc_id] = []
+                hgnc_tx[hgnc_id].append(tx)
+        
+        for i,gene_obj in enumerate(self.hgnc_collection.find({'build': build})):
+            if i > limit:
+                break
+            if add_transcripts:
+                hgnc_id = gene_obj['hgnc_id']
+                tx_objs = hgnc_tx.get(hgnc_id)
+                gene_obj['ens_transcripts'] = tx_objs
+            yield gene_obj
 
-    def nr_genes(self, build=None):
+    def nr_genes(self, build=None, query=None):
         """Return the number of hgnc genes in collection
 
         If build is used, return the number of genes of a certain build
+        
+        Args:
+            build(str): geneome build. '37' or '38'
 
-            Returns:
-                result()
+        Returns:
+            result()
         """
+        query = query or {}
         if build:
-            LOG.info("Fetching all genes from build %s",  build)
+            LOG.debug("Fetching all genes from build %s",  build)
+            query['build'] = build
         else:
-            LOG.info("Fetching all genes")
-
-        return self.hgnc_collection.find({'build':build}).count()
+            LOG.debug("Fetching all genes")
+            
+        
+        nr=0
+        for nr, gene in enumerate(self.hgnc_collection.find(query),1):
+            pass
+        return nr
 
     def drop_genes(self, build=None):
         """Delete the genes collection"""
@@ -217,42 +202,6 @@ class GeneHandler(object):
         else:
             LOG.info("Dropping the hgnc_gene collection")
             self.hgnc_collection.drop()
-
-    def drop_transcripts(self, build=None):
-        """Delete the transcripts collection"""
-        if build:
-            LOG.info("Dropping the transcripts collection, build %s", build)
-            self.transcript_collection.delete_many({'build': build})
-        else:
-            LOG.info("Dropping the transcripts collection")
-            self.transcript_collection.drop()
-
-    def drop_exons(self, build=None):
-        """Delete the exons collection"""
-        if build:
-            LOG.info("Dropping the exons collection, build %s", build)
-            self.exon_collection.delete_many({'build': build})
-        else:
-            LOG.info("Dropping the exons collection")
-            self.exon_collection.drop()
-
-    def ensembl_transcripts(self, build='37'):
-        """Return a dictionary with ensembl ids as keys and transcripts as value.
-
-        Args:
-            build(str)
-
-        Returns:
-            ensembl_transcripts(dict): {<enst_id>: transcripts_obj, ...}
-        """
-        ensembl_transcripts = {}
-        LOG.info("Fetching all transcripts")
-        for transcript_obj in self.transcript_collection.find({'build':build}):
-            enst_id = transcript_obj['transcript_id']
-            ensembl_transcripts[enst_id] = transcript_obj
-        LOG.info("Ensembl transcripts fetched")
-
-        return ensembl_transcripts
 
     def hgncid_to_gene(self, build='37', genes=None):
         """Return a dictionary with hgnc_id as key and gene_obj as value
@@ -368,134 +317,32 @@ class GeneHandler(object):
 
         return alias_genes
 
-    def get_id_transcripts(self, hgnc_id, build='37'):
-        """Return a set with identifier transcript(s)
-
-        Choose all refseq transcripts with NM symbols, if none where found choose ONE with NR,
-        if no NR choose ONE with XM. If there are no RefSeq transcripts identifiers choose the
-        longest ensembl transcript.
-
-        Args:
-            hgnc_id(int)
-            build(str)
-
-        Returns:
-            identifier_transcripts(set)
-
-        """
-        transcripts = self.transcripts(build=build, hgnc_id=hgnc_id)
-
-        identifier_transcripts = set()
-        longest = None
-        nr = []
-        xm = []
-        for tx in transcripts:
-            enst_id = tx['transcript_id']
-            # Should we not check if it is longest?
-            if not longest:
-                longest = enst_id
-            refseq_id = tx.get('refseq_id')
-            if not refseq_id:
-                continue
-
-            if 'NM' in refseq_id:
-                identifier_transcripts.add(enst_id)
-            elif 'NR' in refseq_id:
-                nr.append(enst_id)
-            elif 'XM' in refseq_id:
-                xm.append(enst_id)
-
-        if identifier_transcripts:
-            return identifier_transcripts
-
-        if nr:
-            return set([nr[0]])
-
-        if xm:
-            return set([xm[0]])
-
-        return set([longest])
-
-    def transcripts_by_gene(self, build='37'):
-        """Return a dictionary with hgnc_id as keys and a list of transcripts as value
-
-        Args:
-            build(str)
-
-        Returns:
-            hgnc_transcripts(dict)
-
-        """
-        hgnc_transcripts = {}
-        LOG.info("Fetching all transcripts")
-        for transcript in self.transcript_collection.find({'build':build}):
-            hgnc_id = transcript['hgnc_id']
-            if not hgnc_id in hgnc_transcripts:
-                hgnc_transcripts[hgnc_id] = []
-
-            hgnc_transcripts[hgnc_id].append(transcript)
-
-        return hgnc_transcripts
-
-    def id_transcripts_by_gene(self, build='37'):
-        """Return a dictionary with hgnc_id as keys and a set of id transcripts as value
-
-        Args:
-            build(str)
-
-        Returns:
-            hgnc_id_transcripts(dict)
-        """
-        hgnc_id_transcripts = {}
-        LOG.info("Fetching all id transcripts")
-        for gene_obj in self.hgnc_collection.find({'build': build}):
-            hgnc_id = gene_obj['hgnc_id']
-            id_transcripts = self.get_id_transcripts(hgnc_id=hgnc_id, build=build)
-            hgnc_id_transcripts[hgnc_id] = id_transcripts
-
-        return hgnc_id_transcripts
-
-    def ensembl_genes(self, build='37'):
+    def ensembl_genes(self, build=None, add_transcripts=False, id_transcripts=False):
         """Return a dictionary with ensembl ids as keys and gene objects as value.
 
         Args:
             build(str)
+            transcripts(bool): If transcripts should be included
 
         Returns:
             genes(dict): {<ensg_id>: gene_obj, ...}
         """
+        build = build or '37'
         genes = {}
+        if id_transcripts:
+            add_transcripts = True
 
-        LOG.info("Fetching all genes")
-        for gene_obj in self.hgnc_collection.find({'build':build}):
+        for gene_obj in self.all_genes(build=build, add_transcripts=add_transcripts):
             ensg_id = gene_obj['ensembl_id']
             hgnc_id = gene_obj['hgnc_id']
-
+            transcript_objs = gene_obj.get('ens_transcripts')
+            if id_transcripts and transcript_objs:
+                gene_obj['id_transcripts'] = self.get_id_transcripts(build=build, transcripts=transcript_objs)
             genes[ensg_id] = gene_obj
 
         LOG.info("Ensembl genes fetched")
 
         return genes
-
-    def transcripts(self, build='37', hgnc_id=None):
-        """Return all transcripts.
-
-            If a gene is specified return all transcripts for the gene
-
-        Args:
-            build(str)
-            hgnc_id(int)
-
-        Returns:
-            iterable(transcript)
-        """
-
-        query = {'build': build}
-        if hgnc_id:
-            query['hgnc_id'] = hgnc_id
-
-        return self.transcript_collection.find(query)
-
 
     def to_hgnc(self, hgnc_alias, build='37'):
         """Check if a hgnc symbol is an alias
@@ -586,38 +433,3 @@ class GeneHandler(object):
             intervals[chrom].addi(start, end, i)
 
         return intervals
-
-    def load_exons(self, exons, genes=None, build='37'):
-        """Create exon objects and insert them into the database
-
-        Args:
-            exons(iterable(dict))
-        """
-        genes = genes or self.ensembl_genes(build)
-        for exon in exons:
-            exon_obj = build_exon(exon, genes)
-            if not exon_obj:
-                continue
-
-            res = self.exon_collection.insert_one(exon_obj)
-
-    def exons(self, hgnc_id=None, transcript_id=None,  build=None):
-        """Return all exons
-
-        Args:
-            hgnc_id(int)
-            transcript_id(str)
-            build(str)
-
-        Returns:
-            exons(iterable(dict))
-        """
-        query = {}
-        if build:
-            query['build'] = build
-        if hgnc_id:
-            query['hgnc_id'] = hgnc_id
-        if transcript_id:
-            query['transcript_id'] = transcript_id
-
-        return self.exon_collection.find(query)