+++ tests/adapter/mongo/test_matchmaker.py
@@ -1,30 +1,36 @@
 
-def test_case_mme_update(populated_database, case_obj, user_obj, mme_patient, mme_submission):
+def test_case_mme_update(adapter, case_obj, user_obj, mme_patient, institute_obj, mme_submission):
     """ Test the function that registers an effected individual as submitted to MatchMaker """
-    adapter = populated_database
+    adapter.case_collection.insert_one(case_obj)
+    adapter.institute_collection.insert_one(institute_obj)
+    adapter.user_collection.insert_one(user_obj)
 
     mme_submission['server_responses'] = [ {'patient': mme_patient }]
     
-    # Make sure no case has MME submission:
-    assert adapter.case_collection.find({'mme_submission' : { '$exists' : True}}).count() == 0
-
+    ## GIVEN a database without cases with MME submission:
+    adapter.case_collection.find_one({'mme_submission' : { '$exists' : True}}) is None
 
     updated_case = adapter.case_mme_update(case_obj, user_obj, mme_submission)
 
     # One case has MME submission now
     assert updated_case['mme_submission']
-    assert adapter.case_collection.find({'mme_submission' : { '$exists' : True}}).count()
+    assert adapter.case_collection.find_one({'mme_submission' : { '$exists' : True}})
 
 
-def test_case_mme_delete(populated_database, case_obj, user_obj, mme_patient):
+def test_case_mme_delete(adapter, case_obj, user_obj, institute_obj, mme_patient):
     """ Test the function that updates a case by deleting a MME submission associated to it """
+    adapter.case_collection.insert_one(case_obj)
+    adapter.institute_collection.insert_one(institute_obj)
+    adapter.user_collection.insert_one(user_obj)
 
-    adapter = populated_database
 
     mme_subm_obj = 'mock_submission_object'
 
     # Register a MME submission for a case
-    adapter.case_collection.update_one({ '_id': case_obj['_id']}, {'$set' : {'mme_submission' : mme_subm_obj} })
+    adapter.case_collection.update_one(
+        { '_id': case_obj['_id']}, 
+        {'$set' : {'mme_submission' : mme_subm_obj} }
+    )
     submitted_case = adapter.case_collection.find_one({'mme_submission': {'$exists': True}})
 
     assert submitted_case['mme_submission'] == mme_subm_obj