+++ CHANGELOG.md
@@ -4,11 +4,26 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
-
 ## [4.x.x]
 
 ### Added
 - Picture overview of ideogram, UPDs and ROHs in case-page
+- Show earlier ACMG classification in the variant list
+
+### Fixed
+
+
+## [4.7.1]
+
+### Added
+
+### Fixed
+- Fixed bug on genes page
+
+
+## [4.7.0]
+
+### Added
 - Export genes and gene panels in build GRCh38
 - Search for cases with variants pinned or marked causative in a given gene.
 - Search for cases phenotypically similar to a case also from WUI.
@@ -20,11 +35,15 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Assign case to user when user activates it or asks for rerun
 - Case becomes inactive when it has no assignees
 - Fetch refseq version from entrez and use it in clinvar form
+- Load and export of exons for all genes, independent on refseq
+- Documentation for loading/updating exons
 - Showing SV variant annotations: SV cgh frequencies, gnomad-SV, local SV frequencies
 - Showing transcripts mapping score in segmental duplications
 - Handle requests to Ensembl Rest API  
 - Handle requests to Ensembl Rest Biomart
 - STR variants view now displays GT and IGV link.
+- Description field for gene panels
+- Export exons in build 37 and 38 using the command line
 
 ### Fixed
 - Fixes of and induced by build tests
@@ -33,6 +52,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - MT report only shows variants occurring in the specific individual of the excel sheet
 - Disable SSL certifcate verification in requests to chanjo
 - Updates how intervaltree and pymongo is used to void deprecated functions
+- Increased size of IGV sample tracks
+- Optimized tests
 
 
 ## [4.6.1]